distance determined by chop timing

+++ CHOP JUMP/Assets/Scripts/MovingObject.cs
@@ -18,7 +18,7 @@ public abstract class MovingObject : MonoBehaviour {
 		inverseMoveTime = 1f / moveTime;
 	}
 
-	protected bool Move (int xDir, int yDir, out RaycastHit2D hit) {
+	protected bool Move (float xDir, float yDir, out RaycastHit2D hit) {
 		Vector2 start = transform.position;
 		Vector2 end = start + new Vector2 (xDir, yDir);
 
@@ -34,7 +34,7 @@ public abstract class MovingObject : MonoBehaviour {
 		}
 	}
 
-	protected virtual void AttemptMove <T> (int xDir, int yDir) where T : Component {
+	protected virtual void AttemptMove <T> (float xDir, float yDir) where T : Component {
 		RaycastHit2D hit;
 		bool canMove = Move (xDir, yDir, out hit);
